Module libp2p_networking::network

Network logic

Re-exports

pub use self::error::NetworkError;

Modules

behaviours

networking behaviours wrapping libp2p’s behaviours

def 🔒

defines the swarm and network definition (internal)

error

libp2p network errors Contains the NetworkError snafu types

network_node_handle_error

Re-exports of the snafu errors that NetworkNodeHandleError can throw

node 🔒

functionality of a libp2p network node

Structs

MeshParams

NOTE: mesh_outbound_min <= mesh_n_low <= mesh_n <= mesh_n_high NOTE: mesh_outbound_min <= self.config.mesh_n / 2 parameters fed into gossipsub controlling the structure of the mesh

NetworkDef

Overarching network behaviour performing:

NetworkNode

Network definition

NetworkNodeConfig

describe the configuration of the network

NetworkNodeConfigBuilder

Builder for NetworkNodeConfig.

NetworkNodeHandle

A handle containing:

Enums

ClientRequest

Actions to send from the client to the swarm

NetworkEvent

events generated by the swarm that we wish to relay to the client

NetworkEventInternal

internal representation of the network events only used for event processing before relaying to client

NetworkNodeConfigBuilderError

Error type for NetworkNodeConfigBuilder

NetworkNodeHandleError

Error wrapper type for interacting with swarm handle

NetworkNodeType

this is mostly to estimate how many network connections a node should allow

Functions

deserialize_msg

Deserialize an arbitrary message

gen_multiaddr

Bind all interfaces on port port NOTE we may want something more general in the fture.

gen_transport

Generate authenticated transport

get_random_handle

Given a slice of handles assumed to be larger than 0, chooses one

serialize_msg

Serialize an arbitrary message

spin_up_swarm

a single node, connects them to each other and waits for connections to propagate to all nodes.

Type Aliases

BoxedTransport 🔒

BoxedTransport is a type alias for a boxed tuple containing a PeerId and a StreamMuxerBox.