Struct libp2p_networking::network::behaviours::dht::DHTBehaviour

pub struct DHTBehaviour {

    pub begin_bootstrap: bool,
    pub in_progress_get_closest_peers: HashMap<QueryId, Sender<()>>,
    pub bootstrap_nodes: HashMap<PeerId, HashSet<Multiaddr>>,
    pub event_queue: Vec<DHTEvent>,
    in_progress_get_record_queries: HashMap<QueryId, KadGetQuery>,
    in_progress_put_record_queries: HashMap<QueryId, KadPutQuery>,
    queued_get_record_queries: VecDeque<KadGetQuery>,
    queued_put_record_queries: VecDeque<KadPutQuery>,
    pub kadem: Kademlia<MemoryStore>,
    pub bootstrap_state: Bootstrap,
    pub random_walk: RandomWalk,
    pub peer_id: PeerId,
    pub replication_factor: NonZeroUsize,

}

Behaviour wrapping libp2pโs kademlia included:

publishing API

Request API

bootstrapping into the network

peer discovery

Fields
begin_bootstrap: bool

client approval to begin bootstrap

in_progress_get_closest_peers: HashMap<QueryId, Sender<()>>

in progress queries for nearby peers

bootstrap_nodes: HashMap<PeerId, HashSet<Multiaddr>>

bootstrap nodes

event_queue: Vec<DHTEvent>

List of kademlia events

in_progress_get_record_queries: HashMap<QueryId, KadGetQuery>

List of in-progress get requests

in_progress_put_record_queries: HashMap<QueryId, KadPutQuery>

List of in-progress put requests

queued_get_record_queries: VecDeque<KadGetQuery>

List of previously failled get requests

queued_put_record_queries: VecDeque<KadPutQuery>

List of previously failled put requests

kadem: Kademlia<MemoryStore>

Kademlia behaviour

bootstrap_state: Bootstrap

State of bootstrapping

random_walk: RandomWalk

State of last random walk

peer_id: PeerId

the peer id (useful only for debugging right now)

replication_factor: NonZeroUsize

replication factor

Implementations

impl DHTBehaviour

pub fn begin_bootstrap(&mut self)

Begin the bootstrap process

pub fn query_closest_peers(&mut self, random_peer: PeerId)

Start a query for the closest peers

pub fn new(
kadem: Kademlia<MemoryStore>,
pid: PeerId,
replication_factor: NonZeroUsize
) -> Self

Create a new DHT behaviour

pub fn lookup_peer(&mut self, peer_id: PeerId, chan: Sender<()>)

query a peer (e.g. obtain its address if it exists)

pub fn print_routing_table(&mut self)

print out the routing table to stderr

pub fn add_address(&mut self, peer_id: &PeerId, addr: Multiaddr)

Passthru to kademlia Associate address with kademlia peer

pub fn add_bootstrap_nodes(&mut self, nodes: HashMap<PeerId, HashSet<Multiaddr>>)

Save in case kademlia forgets about bootstrap nodes

pub fn put_record(&mut self, query: KadPutQuery)

Publish a key/value to the kv store. Once replicated upon all nodes, the caller is notified over chan. If there is an error, a crate::network::error::DHTError is sent instead.

pub fn get_record(
&mut self,
key: Vec<u8>,
chan: Sender<Vec<u8>>,
factor: NonZeroUsize,
backoff: ExponentialBackoff,
retry_count: u8
)

Retrieve a value for a key from the DHT. Value (serialized) is sent over chan, and if a value is not found, a crate::network::error::DHTError is sent instead. NOTE: noop if retry_count is 0

fn handle_get_query(
&mut self,
record_results: GetRecordResult,
id: QueryId,
last: bool
)

update state based on recv-ed get query

fn handle_put_query(&mut self, record_results: PutRecordResult, id: QueryId)

Update state based on put query

impl DHTBehaviour

fn dht_handle_event(&mut self, event: KademliaEvent)