Enum libp2p_networking::network::ClientRequest

pub enum ClientRequest {

    BeginBootstrap,
    Shutdown,
    GossipMsg(String, Vec<u8>),
    Subscribe(String, Option<Sender<()>>),
    Unsubscribe(String, Sender<()>),
    DirectRequest {
        pid: PeerId,
        contents: Vec<u8>,
        retry_count: u8,
    },
    DirectResponse(ResponseChannel<DirectMessageResponse>, Vec<u8>),
    Prune(PeerId),
    AddKnownPeers(Vec<(Option<PeerId>, Multiaddr)>),
    IgnorePeers(Vec<PeerId>),
    PutDHT {
        key: Vec<u8>,
        value: Vec<u8>,
        notify: Sender<()>,
    },
    GetDHT {
        key: Vec<u8>,
        notify: Sender<Vec<u8>>,
        retry_count: u8,
    },
    GetConnectedPeerNum(Sender<usize>),
    GetConnectedPeers(Sender<HashSet<PeerId>>),
    GetRoutingTable(Sender<()>),
    LookupPeer(PeerId, Sender<()>),

}

Actions to send from the client to the swarm

Variants

BeginBootstrap

Start the bootstrap process to kademlia

Shutdown

kill the swarm

GossipMsg(String, Vec<u8>)

broadcast a serialized message

Subscribe(String, Option<Sender<()>>)

subscribe to a topic

Unsubscribe(String, Sender<()>)

unsubscribe from a topic

DirectRequest

Fields

pid: PeerId

peer id

contents: Vec<u8>

msg contents

retry_count: u8

number of retries

client request to send a direct serialized message

DirectResponse(ResponseChannel<DirectMessageResponse>, Vec<u8>)

client request to send a direct reply to a message

Prune(PeerId)

prune a peer

AddKnownPeers(Vec<(Option<PeerId>, Multiaddr)>)

add vec of known peers or addresses

IgnorePeers(Vec<PeerId>)

Ignore peers. Only here for debugging purposes. Allows us to have nodes that are never pruned

PutDHT

Fields

key: Vec<u8>

Key to publish under

value: Vec<u8>

Value to publish under

notify: Sender<()>

Channel to notify caller of result of publishing

Put(Key, Value) into DHT relay success back on channel

GetDHT

Fields

key: Vec<u8>

Key to search for

notify: Sender<Vec<u8>>

Channel to notify caller of value (or failure to find value)

retry_count: u8

number of retries to make

Get(Key, Chan)

GetConnectedPeerNum(Sender<usize>)

Request the number of connected peers

GetConnectedPeers(Sender<HashSet<PeerId>>)

Request the set of connected peers

GetRoutingTable(Sender<()>)

Print the routing table to stderr, debugging only

LookupPeer(PeerId, Sender<()>)

Get address of peer