Module libp2p_networking::network::node

functionality of a libp2p network node

Re-exports

pub use self::config::MeshParams;

pub use self::config::NetworkNodeConfig;

pub use self::config::NetworkNodeConfigBuilder;

pub use self::config::NetworkNodeConfigBuilderError;

pub use self::handle::network_node_handle_error;

pub use self::handle::spawn_network_node;

pub use self::handle::NetworkNodeHandle;

pub use self::handle::NetworkNodeHandleError;

pub use self::handle::NetworkNodeReceiver;

Modules

config 🔒

configuration for the libp2p network (e.g. how it should be built)

handle 🔒

libp2p network handle allows for control over the libp2p network

Structs

NetworkNode

Network definition

Constants

ESTABLISHED_LIMIT

Wrapped num of connections

ESTABLISHED_LIMIT_UNWR

Number of connections to a single peer before logging an error

MAX_GOSSIP_MSG_SIZE

Maximum size of a message

_DERIVE_core_fmt_Debug_FOR_NetworkNode 🔒