Struct libp2p_networking::network::node::handle::NetworkNodeHandle

pub struct NetworkNodeHandle<S> {
    network_config: NetworkNodeConfig,
    state: Arc<SubscribableMutex<S>>,
    send_network: UnboundedSender<ClientRequest>,
    listen_addr: Multiaddr,
    peer_id: PeerId,
    id: usize,
    webui_listeners: Arc<Mutex<Vec<Sender<()>>>>,
    receiver: NetworkNodeReceiver,
}

A handle containing:

A reference to the state

Controls for the swarm

Fields
network_config: NetworkNodeConfig

network configuration

state: Arc<SubscribableMutex<S>>

the state of the replica

send_network: UnboundedSender<ClientRequest>

send an action to the networkbehaviour

listen_addr: Multiaddr

the local address we’re listening on

peer_id: PeerId

the peer id of the networkbehaviour

id: usize

human readable id

webui_listeners: Arc<Mutex<Vec<Sender<()>>>>

A list of webui listeners that are listening for changes on this node

receiver: NetworkNodeReceiver
Implementations

impl<S: Default + Debug> NetworkNodeHandle<S>

pub async fn new( config: NetworkNodeConfig, id: usize ) -> Result<Self, NetworkNodeHandleError>

constructs a new node listening on known_addr

pub async fn spawn_handler<F, RET>( self: &Arc<Self>, cb: F ) -> impl Future<Output = ()>where F: Fn(NetworkEvent, Arc<NetworkNodeHandle<S>>) -> RET + Sync + Send + 'static, RET: Future<Output = Result<(), NetworkNodeHandleError>> + Send + 'static, S: Send + 'static,

Spawn a handler F that will be notified every time a new NetworkEvent arrives.

Panics

Will panic if a handler is already spawned

pub async fn wait_to_connect( &self, num_peers: usize, node_id: usize, timeout: Duration ) -> Result<(), NetworkNodeHandleError>where S: Default + Debug,

Wait until at least num_peers have connected, or until timeout time has passed.

Errors

Will return any networking error encountered, or ConnectTimeout if the timeout has elapsed.

pub fn receiver(&self) -> &NetworkNodeReceiver

Receives a reference of the internal NetworkNodeReceiver, which can be used to query for incoming messages.

pub async fn shutdown(&self) -> Result<(), NetworkNodeHandleError>

Cleanly shuts down a swarm node This is done by sending a message to the swarm event handler to stop handling events and a message to the swarm itself to spin down

pub async fn begin_bootstrap(&self) -> Result<(), NetworkNodeHandleError>

Notify the network to begin the bootstrap process

Errors

If unable to send via send_network. This should only happen if the network is shut down.

pub fn listen_addr(&self) -> Multiaddr

Get a reference to the network node handle’s listen addr.

impl<S> NetworkNodeHandle<S>

pub async fn print_routing_table(&self) -> Result<(), NetworkNodeHandleError>

Print out the routing table used by kademlia NOTE: only for debugging purposes currently

Errors

if the client has stopped listening for a response

pub async fn lookup_pid( &self, peer_id: PeerId ) -> Result<(), NetworkNodeHandleError>

Look up a peer’s addresses in kademlia NOTE: this should always be called before any request_response is initiated

Errors

if the client has stopped listening for a response

pub async fn put_record( &self, key: &impl Serialize, value: &impl Serialize ) -> Result<(), NetworkNodeHandleError>

Insert a record into the kademlia DHT

Errors

Will return NetworkNodeHandleError::DHTError when encountering an error putting to DHT

Will return NetworkNodeHandleError::SerializationError when unable to serialize the key or value

pub async fn get_record<V: for<'a> Deserialize<'a>>( &self, key: &impl Serialize, retry_count: u8 ) -> Result<V, NetworkNodeHandleError>

Receive a record from the kademlia DHT if it exists. Must be replicated on at least 2 nodes

Errors

Will return NetworkNodeHandleError::DHTError when encountering an error putting to DHT

Will return NetworkNodeHandleError::SerializationError when unable to serialize the key

Will return NetworkNodeHandleError::DeserializationError when unable to deserialize the returned value

pub async fn get_record_timeout<V: for<'a> Deserialize<'a>>( &self, key: &impl Serialize, timeout: Duration ) -> Result<V, NetworkNodeHandleError>

Get a record from the kademlia DHT with a timeout

Errors

Will return NetworkNodeHandleError::DHTError when encountering an error putting to DHT

Will return NetworkNodeHandleError::TimeoutError when times out

Will return NetworkNodeHandleError::SerializationError when unable to serialize the key

Will return NetworkNodeHandleError::DeserializationError when unable to deserialize the returned value

pub async fn put_record_timeout( &self, key: &impl Serialize, value: &impl Serialize, timeout: Duration ) -> Result<(), NetworkNodeHandleError>

Insert a record into the kademlia DHT with a timeout

Errors

Will return NetworkNodeHandleError::DHTError when encountering an error putting to DHT

Will return NetworkNodeHandleError::TimeoutError when times out

Will return NetworkNodeHandleError::SerializationError when unable to serialize the key or value

Will return NetworkNodeHandleError::SendError when underlying NetworkNode has been killed

pub async fn notify_webui(&self)

Notify the webui that either the state or connection_state has changed.

If the webui is not started, this will do nothing.

Errors

Will return NetworkNodeHandleError::SendError when underlying NetworkNode has been killed

pub async fn subscribe( &self, topic: String ) -> Result<(), NetworkNodeHandleError>

Subscribe to a topic

Errors

Will return NetworkNodeHandleError::SendError when underlying NetworkNode has been killed

pub async fn unsubscribe( &self, topic: String ) -> Result<(), NetworkNodeHandleError>

Unsubscribe from a topic

Errors

Will return NetworkNodeHandleError::SendError when underlying NetworkNode has been killed

pub async fn ignore_peers( &self, peers: Vec<PeerId> ) -> Result<(), NetworkNodeHandleError>

Ignore peers when pruning e.g. maintain their connection

Errors

Will return NetworkNodeHandleError::SendError when underlying NetworkNode has been killed

pub async fn direct_request( &self, pid: PeerId, msg: &impl Serialize ) -> Result<(), NetworkNodeHandleError>

Make a direct request to peer_id containing msg

Errors

Will return NetworkNodeHandleError::SendError when underlying NetworkNode has been killed

Will return NetworkNodeHandleError::SerializationError when unable to serialize msg

pub async fn direct_response( &self, chan: ResponseChannel<DirectMessageResponse>, msg: &impl Serialize ) -> Result<(), NetworkNodeHandleError>

Reply with msg to a request over chan

Errors

Will return NetworkNodeHandleError::SendError when underlying NetworkNode has been killed

Will return NetworkNodeHandleError::SerializationError when unable to serialize msg

pub async fn prune_peer(&self, pid: PeerId) -> Result<(), NetworkNodeHandleError>

Forcefully disconnet from a peer

Errors

If the channel is closed somehow Shouldnt’ happen.

Panics

If channel errors out shouldn’t happen.

pub async fn gossip( &self, topic: String, msg: &impl Serialize ) -> Result<(), NetworkNodeHandleError>

Gossip a message to peers

Errors

Will return NetworkNodeHandleError::SendError when underlying NetworkNode has been killed

Will return NetworkNodeHandleError::SerializationError when unable to serialize msg

pub async fn add_known_peers( &self, known_peers: Vec<(Option<PeerId>, Multiaddr)> ) -> Result<(), NetworkNodeHandleError>

Tell libp2p about known network nodes

Errors

Will return NetworkNodeHandleError::SendError when underlying NetworkNode has been killed

async fn send_request( &self, req: ClientRequest ) -> Result<(), NetworkNodeHandleError>

Send a client request to the network

Errors

Will return NetworkNodeHandleError::SendError when underlying NetworkNode has been killed

pub async fn num_connected(&self) -> Result<usize, NetworkNodeHandleError>

Returns number of peers this node is connected to

Errors

If the channel is closed somehow Shouldnt’ happen.

Panics

If channel errors out shouldn’t happen.

pub async fn connected_pids( &self ) -> Result<HashSet<PeerId>, NetworkNodeHandleError>

return hashset of PIDs this node is connected to

Errors

If the channel is closed somehow Shouldnt’ happen.

Panics

If channel errors out shouldn’t happen.

pub fn id(&self) -> usize

Get a reference to the network node handle’s id.

pub fn peer_id(&self) -> PeerId

Get a reference to the network node handle’s peer id.

pub fn config(&self) -> &NetworkNodeConfig

Return a reference to the network config

pub async fn modify_state<F>(&self, cb: F)where F: FnMut(&mut S),

Modify the state. This will automatically call state_changed and notify_webui

pub fn is_killed(&self) -> bool

Returns true if the network state is killed

pub async fn register_webui_listener(&self) -> Receiver<()>

Register a webui listener

pub async fn state_wait_timeout_until<F>( &self, timeout: Duration, f: F ) -> Result<(), NetworkNodeHandleError>where F: FnMut(&S) -> bool,

Call wait_timeout_until on the state’s [SubscribableMutex]

Errors

Will throw a NetworkNodeHandleError::TimeoutError error upon timeout

pub fn state_wait_timeout_until_with_trigger<'a, F>( &'a self, timeout: Duration, f: F ) -> Timeout<FuturesOrdered<impl Future<Output = ()> + 'a>>where F: FnMut(&S) -> bool + 'a,

Call wait_timeout_until_with_trigger on the state’s [SubscribableMutex]

pub async fn state_wait_until<F>( &self, f: F ) -> Result<(), NetworkNodeHandleError>where F: FnMut(&S) -> bool,

Call wait_until on the state’s [SubscribableMutex]

Errors

Will throw a NetworkNodeHandleError::TimeoutError error upon timeout

impl<S: Clone> NetworkNodeHandle<S>

pub async fn state(&self) -> S

Get a clone of the internal state