Struct libp2p_networking::network::NetworkDef

pub struct NetworkDef { /* private fields */ }

Overarching network behaviour performing:

network topology discovoery

direct messaging

p2p broadcast

connection management

Implementations

impl NetworkDef

pub fn new(
gossipsub: Gossipsub,
kadem: Kademlia<MemoryStore>,
identify: Identify,
request_response: RequestResponse<DirectMessageCodec>,
pruning_enabled: bool,
ignored_peers: HashSet<PeerId>
) -> NetworkDef

Create a new instance of a NetworkDef

impl NetworkDef

Bootstrap functions

pub fn bootstrap(&mut self) -> Result<(), NetworkError>

Bootstrap the network. Make sure at least 1 peer is known, by registering it with .add_address

Errors

Will return a NoKnownPeers error when no known peers are defined

pub fn is_bootstrapped(&self) -> bool

Returns true if the bootstrap state is finished

pub fn is_discovering_peers(&self) -> bool

Returns true if a peer discovery query is in progress

pub fn should_bootstrap(&self) -> bool

Returns true if the bootstrap state is not started

pub fn connection_data(&self) -> Arc<SubscribableRwLock<ConnectionData>>

Returns a reference to the internal ConnectionData

pub fn set_discovering_peers(&mut self, is_discovering: bool)

Toggle whether or not libp2p is discovering peers NOTE: for internal usage only

impl NetworkDef

Address functions

pub fn add_address(&mut self, peer_id: &PeerId, address: Multiaddr)

Add an address

pub fn add_unknown_address(&mut self, addr: Multiaddr)

Add an unknown address

pub fn iter_unknown_addressess(&self) -> impl Iterator<Item = Multiaddr> + '_

Iter all unknown addresses

impl NetworkDef

Peer functions

pub fn add_connected_peer(&mut self, peer_id: PeerId)

Add a connected peer

pub fn remove_connected_peer(&mut self, peer_id: PeerId)

Remove a connected peer

pub fn connected_peers(&self) -> HashSet<PeerId>

Get a list of the connected peers

pub fn add_known_peer(&mut self, peer_id: PeerId)

Add a known peer

pub fn known_peers(&self) -> HashSet<PeerId>

Get a list of the known peers

pub fn add_connecting_peer(&mut self, a_peer: PeerId)

Add a connecting peer

pub fn remove_peer(&mut self, peer_id: PeerId)

Remove a peer, both connecting and connected

pub fn get_peers(&self) -> HashSet<PeerId>

Get a list of peers, both connecting and connected

pub fn get_peers_to_prune(&self, max_num_peers: usize) -> Vec<PeerId>

Return a list of peers to prune, based on given max_num_peers

pub fn toggle_pruning(&mut self, is_enabled: bool)

Toggle pruning

pub fn query_closest_peers(&mut self, random_peer: PeerId)

Start a query for the closest peers

pub fn extend_ignored_peers(&mut self, peers: Vec<PeerId>)

Add a list of peers to the ignored peers list

impl NetworkDef

Gossip functions

pub fn publish_gossip(&mut self, topic: Topic, contents: Vec<u8>)

Publish a given gossip

pub fn subscribe_gossip(&mut self, t: &str)

Subscribe to a given topic

pub fn unsubscribe_gossip(&mut self, t: &str)

Unsubscribe from a given topic

pub fn drain_publish_gossips(&mut self)

Attempt to drain the internal gossip list, publishing each gossip

pub fn retry_put_dht(&mut self)

attempt to put all unstarted put querys to DHT

impl NetworkDef

DHT functions

pub fn put_record(&mut self, query: KadPutQuery)

Publish a key/value to the kv store. Once replicated upon all nodes, the caller is notified over chan. If there is an error, a [DHTError] is sent instead.

pub fn new_put_uid(&mut self) -> usize

generates new uid for put record query

pub fn get_record(
&mut self,
key: Vec<u8>,
chan: Sender<Vec<u8>>,
factor: NonZeroUsize
)

Retrieve a value for a key from the DHT. Value (serialized) is sent over chan, and if a value is not found, a [DHTError] is sent instead.

impl NetworkDef

Request/response functions

pub fn add_direct_request(&mut self, pid: PeerId, msg: Vec<u8>)

Add a direct request for a given peer

pub fn add_direct_response(
&mut self,
chan: ResponseChannel<DirectMessageResponse>,
msg: Vec<u8>
)

Add a direct response for a channel