Type Alias libp2p_networking::network::BoxedTransport

type BoxedTransport = Boxed<(PeerId, StreamMuxerBox)>;

BoxedTransport is a type alias for a boxed tuple containing a PeerId and a StreamMuxerBox.

This type is used to represent a transport in the libp2p network framework. The PeerId is a unique identifier for each peer in the network, and the StreamMuxerBox is a type of multiplexer that can handle multiple substreams over a single connection.

Aliased Type

struct BoxedTransport {
    inner: Box<dyn Abstract<(PeerId, StreamMuxerBox)> + Unpin + Send>,
}

Fields
inner: Box<dyn Abstract<(PeerId, StreamMuxerBox)> + Unpin + Send>