Struct libp2p_networking::reexport::PeerId

pub struct PeerId {
    multihash: Multihash<64>,
}

Identifier of a peer of the network.

The data is a CIDv0 compatible multihash of the protobuf encoded public key of the peer as specified in specs/peer-ids.

Fields
multihash: Multihash<64>
Implementations

impl PeerId

pub fn from_public_key(key: &PublicKey) -> PeerId

Builds a PeerId from a public key.

pub fn from_bytes(data: &[u8]) -> Result<PeerId, Error>

Parses a PeerId from bytes.

pub fn from_multihash(multihash: Multihash<64>) -> Result<PeerId, Multihash<64>>

Tries to turn a Multihash into a PeerId.

If the multihash does not use a valid hashing algorithm for peer IDs, or the hash value does not satisfy the constraints for a hashed peer ID, it is returned as an Err.

pub fn try_from_multiaddr(address: &Multiaddr) -> Option<PeerId>

Tries to extract a PeerId from the given Multiaddr.

In case the given Multiaddr ends with /p2p/<peer-id>, this function will return the encapsulated PeerId, otherwise it will return None.

pub fn random() -> PeerId

Generates a random peer ID from a cryptographically secure PRNG.

This is useful for randomly walking on a DHT, or for testing purposes.

pub fn to_bytes(&self) -> Vec<u8, Global>

Returns a raw bytes representation of this PeerId.

pub fn to_base58(&self) -> String

Returns a base-58 encoded string of this PeerId.

pub fn is_public_key(&self, public_key: &PublicKey) -> Option<bool>

Checks whether the public key passed as parameter matches the public key of this PeerId.

Returns None if this PeerIds hash algorithm is not supported when encoding the given public key, otherwise Some boolean as the result of an equality check.