Struct libp2p_networking::network::NetworkDef

pub struct NetworkDef {
    gossipsub: GossipBehaviour,
    pub dht: DHTBehaviour,
    identify: IdentifyBehaviour,
    pub request_response: DMBehaviour,
}

Overarching network behaviour performing:

network topology discovoery

direct messaging

p2p broadcast

connection management

Fields
gossipsub: GossipBehaviour

purpose: broadcasting messages to many peers NOTE gossipsub works ONLY for sharing messsages right now in the future it may be able to do peer discovery and routing https://github.com/libp2p/rust-libp2p/issues/2398

dht: DHTBehaviour

purpose: peer routing purpose: storing pub key <-> peer id bijection

identify: IdentifyBehaviour

purpose: identifying the addresses from an outside POV

request_response: DMBehaviour

purpose: directly messaging peer

Implementations

impl NetworkDef

pub fn new(
gossipsub: GossipBehaviour,
dht: DHTBehaviour,
identify: IdentifyBehaviour,
request_response: DMBehaviour
) -> NetworkDef

Create a new instance of a NetworkDef

impl NetworkDef

Address functions

pub fn add_address(&mut self, peer_id: &PeerId, address: Multiaddr)

Add an address

impl NetworkDef

Gossip functions

pub fn publish_gossip(&mut self, topic: Topic, contents: Vec<u8>)

Publish a given gossip

pub fn subscribe_gossip(&mut self, t: &str)

Subscribe to a given topic

pub fn unsubscribe_gossip(&mut self, t: &str)

Unsubscribe from a given topic

impl NetworkDef

DHT functions

pub fn put_record(&mut self, query: KadPutQuery)

Publish a key/value to the kv store. Once replicated upon all nodes, the caller is notified over chan. If there is an error, a super::error::DHTError is sent instead.

pub fn get_record(
&mut self,
key: Vec<u8>,
chan: Sender<Vec<u8>>,
factor: NonZeroUsize,
retry_count: u8
)

Retrieve a value for a key from the DHT. Value (serialized) is sent over chan, and if a value is not found, a super::error::DHTError is sent instead.

impl NetworkDef

Request/response functions

pub fn add_direct_request(
&mut self,
peer_id: PeerId,
data: Vec<u8>,
retry_count: u8
)

Add a direct request for a given peer

pub fn add_direct_response(
&mut self,
chan: ResponseChannel<DirectMessageResponse>,
msg: Vec<u8>
)

Add a direct response for a channel